+++ src/main/java/life/majiang/community/community/controller/QuestionController.java
@@ -1,9 +1,6 @@
 package life.majiang.community.community.controller;
 
 import life.majiang.community.community.dto.QuestionDTO;
-import life.majiang.community.community.mapper.QuestionExtMapper;
-import life.majiang.community.community.mapper.QuestionMapper;
-import life.majiang.community.community.model.Question;
 import life.majiang.community.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -18,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id")Integer id,
+    public String question(@PathVariable(name = "id")Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         //累加阅读数
